requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0.post1

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Blockchain
solana==0.30.2
web3==5.24.0

# HTTP Client
aiohttp==3.8.1
httpx==0.25.2

# Testing
pytest==6.2.5
pytest-asyncio==0.15.1
pytest-cov==2.12.1
pytest-mock==3.12.0

# Code Quality
black==21.7b0
flake8==3.9.2
mypy==0.910
isort==5.12.0
pre-commit==3.5.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0 